Added hermaphroditic parent functionality
Removed 'updateSire', 'updateDam' and 'updatePosterior', replaced with 'updateParentPosterior'.  This function checks if the parent is the sire or dam in each family and calls the appropriate functions ('posteriorSire_new' or 'posteriorDam_new' etc. as appropriate).

This new function allows a single individual to be both a sire and a dam, within the same pedigree.  This is applicable to some trees, which possess male and female flowers.